+++ docxbox/docx/xml/docx_xml.h
@@ -22,8 +22,6 @@ class docx_xml {
   char **argv_;
 
   std::string document_text_;
-
-  bool RemoveNodes(std::vector<tinyxml2::XMLElement*> nodes);
 };
 
 #endif  // DOCXBOX_DOCX_XML_DOCX_XML_H_