+++ docxbox/docx/xml/docx_xml_fields.h
@@ -4,6 +4,7 @@
 #define DOCXBOX_DOCX_XML_DOCX_XML_FIELDS_H_
 
 #include <docxbox/docx/xml/docx_xml.h>
+#include <docxbox/docx/xml/docx_xml_remove.h>
 #include <docxbox/helper/helper_file.h>
 #include <docxbox/helper/helper_string.h>
 
@@ -19,10 +20,10 @@ class docx_xml_fields:docx_xml {
 
   void CollectMergeFields(const std::string& path_xml);
 
-  bool SetFieldValue(
+  bool SetFieldText(
       const std::string& path_xml,
       const std::string &field_identifier,
-      const std::string &value);
+      const std::string &text);
 
   void Output(bool as_json);
  private:
@@ -33,17 +34,7 @@ class docx_xml_fields:docx_xml {
   // Vector of vectors: fields per XML file
   std::vector <std::vector<std::string>> fields_in_xml_files_;
 
-  tinyxml2::XMLElement *previous_run_;
-  //tinyxml2::XMLElement *previous_field_begin_;
-  //tinyxml2::XMLElement *previous_field_end_;
-
-  tinyxml2::XMLElement *run_around_field_begin_;
-  tinyxml2::XMLElement *run_around_field_instrText_;
-  tinyxml2::XMLElement *run_around_field_end_;
-
   bool is_inside_searched_field_ = false;
-
-  std::vector<tinyxml2::XMLElement*> nodes_to_be_removed_;
 
   void CollectFieldsFromNodes(tinyxml2::XMLElement *node);