+++ CHANGELOG.md
@@ -3,6 +3,7 @@ docxBox Changelog
 
 V0.0.3 - Not finished yet
 -------------------------
+* Add set field value (sfv)
 * Ensure re-zipped DOCX is rated valid when opened in MS word
 * Add DOCX check before attempting to extract: must be a ZIP archive
 * Add list files containing search-string (lsl)